+++ CHANGES.md
@@ -2,7 +2,7 @@
 
 ## Next
 
-* Fix temporal type formatting in BigQueryRestriction to consistently use 6 decimal digits
+* PR #258: Fix temporal type formatting in BigQueryRestriction to consistently use 6 decimal digits
 
 ## 1.0.0 - 2025-02-25